Struct datom_bigdecimal::BigDecimal

pub struct BigDecimal { /* private fields */ }

A big decimal type.

Implementations

impl BigDecimal

pub fn new(digits: BigInt, scale: i64) -> BigDecimal

Creates and initializes a BigDecimal.

pub fn parse_bytes(buf: &[u8], radix: u32) -> Option<BigDecimal>

Creates and initializes a BigDecimal.

Decodes using str::from_utf8 and forwards to BigDecimal::from_str_radix. Only base-10 is supported.

Examples

use bigdecimal::{BigDecimal, Zero};

assert_eq!(BigDecimal::parse_bytes(b"0", 10).unwrap(), BigDecimal::zero());
assert_eq!(BigDecimal::parse_bytes(b"13", 10).unwrap(), BigDecimal::from(13));

pub fn with_scale(&self, new_scale: i64) -> BigDecimal

Return a new BigDecimal object equivalent to self, with internal scaling set to the number specified. If the new_scale is lower than the current value (indicating a larger power of 10), digits will be dropped (as precision is lower)

pub fn with_prec(&self, prec: u64) -> BigDecimal

Return a new BigDecimal object with precision set to new value

pub fn sign(&self) -> Sign

Return the sign of the BigDecimal as num::bigint::Sign.

Examples

extern crate num_bigint;
extern crate bigdecimal;
use std::str::FromStr;

assert_eq!(bigdecimal::BigDecimal::from_str("-1").unwrap().sign(), num_bigint::Sign::Minus);
assert_eq!(bigdecimal::BigDecimal::from_str("0").unwrap().sign(), num_bigint::Sign::NoSign);
assert_eq!(bigdecimal::BigDecimal::from_str("1").unwrap().sign(), num_bigint::Sign::Plus);

pub fn as_bigint_and_exponent(&self) -> (BigInt, i64)

Return the internal big integer value and an exponent. Note that a positive exponent indicates a negative power of 10.

Examples

extern crate num_bigint;
extern crate bigdecimal;
use std::str::FromStr;

assert_eq!(bigdecimal::BigDecimal::from_str("1.1").unwrap().as_bigint_and_exponent(),
           (num_bigint::BigInt::from_str("11").unwrap(), 1));

pub fn into_bigint_and_exponent(self) -> (BigInt, i64)

Convert into the internal big integer value and an exponent. Note that a positive exponent indicates a negative power of 10.

Examples

extern crate num_bigint;
extern crate bigdecimal;
use std::str::FromStr;

assert_eq!(bigdecimal::BigDecimal::from_str("1.1").unwrap().into_bigint_and_exponent(),
           (num_bigint::BigInt::from_str("11").unwrap(), 1));

pub fn digits(&self) -> u64

Number of digits in the non-scaled integer representation

pub fn abs(&self) -> BigDecimal

Compute the absolute value of number

pub fn double(&self) -> BigDecimal

pub fn half(&self) -> BigDecimal

Divide this efficiently by 2

Note, if this is odd, the precision will increase by 1, regardless of the context’s limit.

pub fn square(&self) -> BigDecimal

pub fn cube(&self) -> BigDecimal

pub fn sqrt(&self) -> Option<BigDecimal>

Take the square root of the number

If the value is < 0, None is returned

pub fn cbrt(&self) -> BigDecimal

Take the cube root of the number

pub fn inverse(&self) -> BigDecimal

Compute the reciprical of the number: x-1

pub fn round(&self, round_digits: i64) -> BigDecimal

Return number rounded to round_digits precision after the decimal point

pub fn is_integer(&self) -> bool

Return true if this number has zero fractional part (is equal to an integer)

pub fn exp(&self) -> BigDecimal

Evaluate the natural-exponential function ex

pub fn normalized(&self) -> BigDecimal